types(frame handles): AnyValueMatches_155630 gets its REAL signature; the i32-typed frame members become CImage*
The reverse-lookup's own body proved the params ((CImageSet*)val)->FindFrame(
(CImage*)a1,(char*)a2,(int*)a3) - so declare them that way:
AnyValueMatches_155630(CImage* frame, char* outName, i32* outIndex). THE FN
ITSELF WENT 75.9 -> 100.00 EXACT: its '@early-stop regalloc wall' was the fake
i32,i32,i32 signature's codegen, dissolved by the truth.

Members that stored frames as i32/void* are typed CImage*: CSBI_Image::m_30
(inherited by the whole SBI_ImageSet/ImageSetAni/AniPlayer/MenuItem/WellGoo/
WarlordHead family - every '(CImage*)m_30' and '(i32)frame' cast falls out),
CMgrSettings::m_38, CActionOptionsMenuBar::m_button0/1Frame. CSBI_MenuItem::
m_38 'void* (CImageSet | CSprite)' typed CImageSet*.

Two more gated-frame-table views dissolved onto CImageSet: CWhConfig
(SBI_WarlordHead.h - its lo/hi gate comments were even swapped) and the
SbiTabFrame/SbiTabFrameSub hop in SBI_RectOnly (its 'm_14->m_10' was literally
m_frames[4]).

Byte results: every touched fn HOLDS (mgrsettings 89.05 unchanged, all AOMB/
SpriteLoaders/WellGoo/Play legs identical); AnyValueMatches +24. Exact 2388.
Cleanliness vs baseline: numeric casts -52, )m_ -26, void* m_ -13, m_<hex> -84.

(cherry picked from commit 8e99a789c4d1ba3068dcf2d79b71b9775e56fb0d)

+++ src/Gruntz/AniPlayer.cpp
@@ -98,7 +98,7 @@ i32 CAniPlayer::RenderCel_0e5c10() {
     } else {
         cel = 0;
     }
-    m_30 = (i32)cel;
+    m_30 = cel;
     if (cel != 0) {
         i32 surfaceCtx = (i32)g_gameReg->m_world->m_drawTarget->m_backPair;
         cel->RenderFrame(
